requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-autoslug==1.9.9
django-cleanup==8.0.0
django-jazzmin==2.6.0
django-multiupload==0.6.1
djangorestframework==3.14.0
idna==3.4
lyricsgenius==3.0.1
psycopg2==2.9.6
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
spotipy==2.23.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
